First commit
The first commit to the repo with all the latest changes. The commit
does not contain any third partybinary dependencies or sources. This
should be downloaded by hand. I'm thinking about providing a Maven
pom.xml or Ant build.xml to get these dependenices easily, but this
won't happen soon for sure.